src/requirements.txt
flask
flask-socketio
RPi.GPIO==0.7.1
